Move crate-level migration test to module
