+++ crates/sui-genesis-builder/src/stardust/mod.rs
@@ -9,3 +9,5 @@ pub mod migration;
 pub mod native_token;
 pub mod parse;
 pub mod types;
+#[cfg(test)]
+mod migration_tests;
